requirements.txt
# UI
gradio>=4.0.0

# Face detection & swap
insightface>=0.7.3
onnxruntime>=1.16.0        # CPU; swap for onnxruntime-gpu if you have CUDA

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Invisible watermarking
invisible-watermark>=0.2.0

# Deepfake detection
timm>=0.9.0                # backbone models
torch>=2.0.0
torchvision>=0.15.0
